Wed, Oct 19, 2022, 2:29 PM +00:00

+++ readability/readability.c
@@ -17,8 +17,8 @@ int main(void)
     int sentenceCount = countSentences(text, len);
     // organiseToArrays(text, wordCount);
     printf("WordCount: %i, letterCount: %i, sentenceCount: %i\n", wordCount, letterCount, sentenceCount);
-    double L = (letterCount / wordCount * 100);
-    double S = (sentenceCount / wordCount * 100);
+    double L = ((double) letterCount / wordCount * 100);
+    double S = ((double) sentenceCount / wordCount * 100);
     double index = (0.0588 * L) - (0.296 * S) - 15.8;
     printf("%f\n", index);
     if (index > 16)
